requirements.txt
requests
python-dotenv
pushover_complete
tenacity
